A bit of improvement done
