+++ src/components/lists/quadrant/DataLists.tsx
@@ -1,147 +1,20 @@
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
 import React, { useEffect, useState } from 'react';
-import { v4 as uuidv4 } from 'uuid';
-
+import { v4 as uuidv4, v4 } from 'uuid';
 import {
-  Utilities,
   RadarUtilities,
   useRadarState,
   useDataState,
   BlipType,
-  RadarOptionsType
+  HorizonKey,
+  Utilities
 } from '@undp_sdg_ai_lab/undp-radar';
-import { ScrollableDiv } from '../components/ScrollableDiv';
+import { Box, Accordion, AccordionItem } from '@chakra-ui/react';
 
+import { QuadrantDataListItem } from './QuadrantDataListItem';
+import { UuidNameObjectType as UuidNameObjType } from './type';
 import './DataLists.scss';
-import {
-  Accordion,
-  AccordionButton,
-  AccordionIcon,
-  AccordionItem,
-  AccordionPanel,
-  Badge,
-  Box,
-  Button,
-  Flex,
-  Text
-} from '@chakra-ui/react';
 
-type ListMatrixItem = { uuid: string; name: string };
-
-interface Props {
-  quadrant: ListMatrixItem;
-  horizon?: ListMatrixItem | null;
-  blips: BlipType[];
-  setHoveredItem: (payload: BlipType | null) => void;
-  setSelectedItem: (item: BlipType) => void;
-}
-
-const QuadrantItemList: React.FC<Props> = ({
-  quadrant,
-  horizon = null,
-  blips,
-  setHoveredItem,
-  setSelectedItem
-}) => {
-  const {
-    state: {
-      keys: { titleKey, quadrantKey, horizonKey }
-    }
-  } = useDataState();
-
-  const onMouseLeave = () => setHoveredItem(null); // equal for all
-
-  return (
-    <ScrollableDiv maxHeight={400}>
-      <Accordion allowToggle>
-        {blips.map((blip) => {
-          const onMouseEnter = () => setHoveredItem(blip);
-          if (
-            blip[quadrantKey] === quadrant.name &&
-            (horizon === null || blip[horizonKey] === horizon.name)
-          )
-            return (
-              <AccordionItem
-                onMouseEnter={onMouseEnter}
-                onMouseLeave={onMouseLeave}
-                key={blip.id}
-              >
-                <h5>
-                  <AccordionButton>
-                    <Box as='h6' flex='1' textAlign='left'>
-                      {blip[titleKey]}
-                    </Box>
-                    <AccordionIcon />
-                  </AccordionButton>
-
-                  <AccordionPanel pb={4}>
-                    <Box bg={'#EDF2F7'}>
-                      <Flex direction={'column'} minHeight={'200px'} p='5'>
-                        <Box>
-                          <Text mb='2'>Description</Text>
-                          <Text fontWeight={'400'} fontSize={'md'}>
-                            {blip.Description}
-                          </Text>
-                        </Box>
-                        <Flex flexWrap={'wrap'} my='5'>
-                          <Badge
-                            isTruncated
-                            my='1'
-                            mx='1'
-                            variant='subtle'
-                            colorScheme='orange'
-                          >
-                            🌋{' ' + blip['Disaster Cycle']}
-                          </Badge>
-                          <Badge
-                            isTruncated
-                            my='1'
-                            mx='1'
-                            variant='subtle'
-                            colorScheme='green'
-                          >
-                            🏠{' ' + blip['Un Host Organisation']}
-                          </Badge>
-                          <Badge
-                            isTruncated
-                            my='1'
-                            mx='1'
-                            variant='subtle'
-                            colorScheme='purple'
-                          >
-                            📍{' ' + blip['Country of Implementation']}
-                          </Badge>
-                          <Badge
-                            isTruncated
-                            my='1'
-                            mx='1'
-                            variant='subtle'
-                            colorScheme='cyan'
-                          >
-                            🎯{' ' + blip['SDG']}
-                          </Badge>
-                        </Flex>
-                        <Button
-                          onClick={() => setSelectedItem(blip)}
-                          colorScheme='blue'
-                          borderRadius={'0'}
-                        >
-                          More
-                        </Button>
-                      </Flex>
-                    </Box>
-                  </AccordionPanel>
-                </h5>
-              </AccordionItem>
-            );
-          return null;
-        })}
-      </Accordion>
-    </ScrollableDiv>
-  );
-};
-
-type HorizonItemType = { uuid: string; name: string };
+type DisplayBlipsByQuadrantType = Record<string, BlipType[]>;
 
 export const QuadrantDataLists: React.FC = () => {
   const {
@@ -150,67 +23,59 @@ export const QuadrantDataLists: React.FC = () => {
       isFiltered,
       filteredBlips,
       selectedQuadrant: selectedQuad,
-      radarData,
-      hoveredTech,
-      hoveredItem
-    },
-    actions: { setHoveredItem, setSelectedItem }
+      radarData: { horizons }
+    }
   } = useRadarState();
   const {
     state: { keys }
   } = useDataState();
 
-  const [displayBlips, setDisplayBlips] = useState<BlipType[]>([]);
-  const [displayHorizons, setDisplayHorizons] = useState<HorizonItemType[]>([]);
-  // const [];
+  const [displayBlipsByHorizon, setDisplayBlipsByHorizon] =
+    useState<DisplayBlipsByQuadrantType>({});
+
+  const [displayHorizons, setDisplayHorizons] = useState<UuidNameObjType[]>([]);
 
   useEffect(() => {
-    const filtered = isFiltered ? filteredBlips : blips;
-
-    const quads = RadarUtilities.getQuadrants(displayBlips, keys.quadrantKey);
-    if (selectedQuad) {
-      console.log(
-        quads,
-        quads.indexOf(selectedQuad),
-        selectedQuad,
-        filtered[0].quadrantIndex
-      );
-      filtered.filter((b) => b.quadrantIndex === quads.indexOf(selectedQuad));
-    }
-    setDisplayBlips(filtered);
-  }, [blips, isFiltered, filteredBlips, selectedQuad]);
-
-  useEffect(() => {
+    const displayBlips = isFiltered ? filteredBlips : blips;
+    let newDisplayBlipsByHorizon: DisplayBlipsByQuadrantType = {};
     const quads = RadarUtilities.getQuadrants(displayBlips, keys.quadrantKey);
     if (selectedQuad) {
       const newHorizons: string[] = [];
       const indexOfSelectedQuadrant = quads.indexOf(selectedQuad);
       displayBlips.forEach((blip) => {
         if (blip.quadrantIndex === indexOfSelectedQuadrant) {
-          const newHorizon = blip[keys.horizonKey];
+          const newHorizon = blip[keys.horizonKey as HorizonKey] as string;
+          newDisplayBlipsByHorizon = {
+            ...newDisplayBlipsByHorizon,
+            [newHorizon]: [
+              ...(newDisplayBlipsByHorizon[newHorizon] || []),
+              blip
+            ]
+          };
           if (!newHorizons.includes(newHorizon)) newHorizons.push(newHorizon);
         }
       });
-      setDisplayHorizons(newHorizons.map((h) => ({ uuid: uuidv4(), name: h })));
+      setDisplayHorizons(
+        horizons
+          .filter((h) => newHorizons.includes(h))
+          .map((h) => ({ uuid: uuidv4(), name: h }))
+      );
+      setDisplayBlipsByHorizon(newDisplayBlipsByHorizon);
       return;
+    } else {
+      setDisplayHorizons(quads.map((q) => ({ uuid: uuidv4(), name: q })));
+      setDisplayBlipsByHorizon(newDisplayBlipsByHorizon);
     }
-    setDisplayHorizons(quads.map((q) => ({ uuid: uuidv4(), name: q })));
-  }, [displayBlips, keys]);
+  }, [filteredBlips, blips, isFiltered, keys, horizons]);
 
   return (
-    <Box as='section'>
+    <Box>
       <Accordion allowToggle>
-        {displayHorizons.map((horizon) => (
-          <AccordionItem key={horizon.uuid}>
+        {displayHorizons.map((horizon, index) => (
+          <AccordionItem key={index}>
             <QuadrantDataListItem
-              radarData={radarData}
-              hoveredTech={hoveredTech}
-              setHoveredItem={setHoveredItem}
-              hoveredItem={hoveredItem}
-              setSelectedItem={setSelectedItem}
-              blips={blips}
-              headers={displayHorizons}
-              horizon={horizon}
+              horizonName={Utilities.capitalize(horizon.name)}
+              quadrantBlips={displayBlipsByHorizon[horizon.name]}
             />
           </AccordionItem>
         ))}
@@ -373,58 +238,4 @@ export const QuadrantDataLists: React.FC = () => {
   //     )}
   //   </section>
   // );
-};
-
-interface QuadrantDataListItemProps {
-  radarData: RadarOptionsType;
-  horizon: ListMatrixItem;
-  blips: BlipType[];
-  hoveredItem: BlipType | null;
-  hoveredTech: string | null;
-  setHoveredItem: (payload: BlipType | null) => void;
-  setSelectedItem: (item: BlipType) => void;
-  headers: ListMatrixItem[];
-}
-
-const QuadrantDataListItem: React.FC<QuadrantDataListItemProps> = ({
-  radarData,
-  horizon,
-  blips,
-  hoveredItem,
-  hoveredTech,
-  setHoveredItem,
-  setSelectedItem,
-  headers
-}) => {
-  return (
-    <div>
-      {headers.map((header, index) => (
-        <div key={index}>
-          {/*TODO: Usage of uuidv4() causes bug where accordian glitches in height*/}
-          <h5>
-            <AccordionButton>
-              <Box
-                key={`${header.uuid}-${horizon.uuid}`}
-                flex='1'
-                textAlign='left'
-                as='h5'
-              >
-                {Utilities.capitalize(horizon.name)}
-              </Box>
-              <AccordionIcon />
-            </AccordionButton>
-          </h5>
-          <AccordionPanel>
-            <QuadrantItemList
-              blips={blips}
-              quadrant={header}
-              horizon={horizon}
-              setHoveredItem={setHoveredItem}
-              setSelectedItem={setSelectedItem}
-            />
-          </AccordionPanel>
-        </div>
-      ))}
-    </div>
-  );
 };
