+++ src/components/lists/quadrant/type.ts
@@ -1,0 +1,1 @@
+export type UuidNameObjectType = { uuid: string; name: string };
